# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fancy RSVP API",
    "description": "Backend API for Fancy RSVP event management and guest RSVP tracking platform",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Version 1 API Base"
    }
  ],
  "paths": {
...
    "/events": {
      "get": {
        "summary": "List events for authenticated organizer",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": { "description": "List of events" }
        }
      },
      "post": {
        "summary": "Create a new event",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["slug", "templateType", "title", "eventDate"],
                "properties": {
                  "slug": { "type": "string" },
                  "templateType": { "type": "string" },
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "eventDate": { "type": "string", "format": "date-time" },
                  "locationName": { "type": "string" },
                  "locationAddress": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Event created" }
        }
      }
    },
...
    "/events/{eventId}/rsvps": {
      "get": {
        "summary": "List RSVPs for event",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "eventId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } }
        ],
        "responses": {
          "200": { "description": "List of RSVPs" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}